implementing support for Grafana Queries global variables and adding support for macros in Grafana.
